pyproject.toml
[project]
name = "open-vocabulary-od-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "onnx>=1.18.0",
    "onnxruntime>=1.22.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "seaborn>=0.13.2",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "transformers>=4.52.4",
    "ultralytics>=8.3.159",
]
